requirements.txt
alembic==1.4.2
aniso8601==7.0.0
click==7.1.1
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
promise==2.3
ptvsd==4.2.8
PyJWT==1.4.2
python-dateutil==2.8.1
python-editor==1.0.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
